[HOTFIX] Memory leak
So, a simple fix never ends up simple, does it?

The theory here was that the read layer created an issue because it used to
much memory on reading large tables. This needed som refactoring anyways,
so I decided to reimplement it with a different approach, using a Dataset
interface as an abstraction, and using channels to stream data from the
database to the json writer in the frontend layer.

Using channels for this is well within "correct" Go, and in this case also
provides proper back pressure.

Unfortunately it turned out that the problem happens when data is written
into a post layer with a table missing form the database. Somehow the error
this caused was not handled properly, and it kept eating data until it got
fat and blew up.

I was already done with the read refactoring, so decided to also refactor
the write layer in the same way. It is now set up in a publisher->consumer
setup where the consumer is consuming the input json stream and publishing
it to the dataset write through a channel as well. A bit off extra effort was
needed here to control the flow, specifically around the publisher terminating
(from an error) before the consumer was finnished reading from the input.

Also spent time cleaning out the (non functional) CDC code